Refactored clientThread. Clientthread now takes in datastructures for messages, users, etc. Any changes to these data structures will be seen in the other threads.
Handling of requests is now more readable.

+++ src/main/java/CardGame/Requests/RequestGetMessages.java
@@ -9,19 +9,19 @@ package CardGame.Requests;
  * Created by tom on 11/03/17.
  */
 public class RequestGetMessages extends RequestProtocol {
-    private long offset;
+    private int offset;
 
-    public RequestGetMessages(int protocolId, int type, long offset) {
+    public RequestGetMessages(int protocolId, int type, int offset) {
         super(protocolId, type);
         this.offset = offset;
     }
 
-    public RequestGetMessages(int type, long offset) {
+    public RequestGetMessages(int type, int offset) {
         super(type);
         this.offset = offset;
     }
 
-    public long getOffset() {
+    public int getOffset() {
         return offset;
     }
